+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+## [0.1.2]
+
+- Fixed function call highlighting
+- Changed error highlighting color
+
 ## [0.1.1]
 
 - Prettified structure view
@@ -14,14 +19,12 @@
 - Fixed some highlighting and formatting issues for Leo
 
 ## [0.0.2]
-
 - Added basic code completion
 - Added basic formatting rules
 - Implemented structure view
 - Implemented go to symbol
 
 ## [0.0.1-3]
-
 - Added code folding support
 - Added parse recovering
 - Reference circuit components in circuit expressions